Resolve feedback from Nikita Lipsky
