+++ benchmarks/SkikoBenchmarks/compare_benchmarks.main.kts
@@ -1,6 +1,6 @@
 #!/usr/bin/env kotlin
 
-import java.io.File
+@file:Import("benchmark_project.main.kts")
 
 /**
  * Usage:
@@ -356,20 +356,6 @@ fun allBenchmarkNames(): Set<String> {
     }
 
     return names
-}
-
-fun findBenchmarkProjectDir(): File {
-    val current = File(".").canonicalFile
-    if (current.resolve("build.gradle.kts").exists() && current.name == "SkikoBenchmarks") {
-        return current
-    }
-
-    val fromRepoRoot = current.resolve("benchmarks/SkikoBenchmarks")
-    if (fromRepoRoot.resolve("build.gradle.kts").exists()) {
-        return fromRepoRoot.canonicalFile
-    }
-
-    throw IllegalStateException("Cannot locate benchmarks/SkikoBenchmarks project")
 }
 
 main(args)
